<udev: Add a udev rules that should benefit for systemd based system>
This allow use of Mooltipass Mini with systemd based computers, it
benefit from uaccess tag to allow logged user to access the device.
In addition the rule file include the rules needed for libusb to
access the device too.

This could be one step to merge the provided udev rules for both, at
least, Arch Linux and Ubuntu (16.04).