Add service exists api
New API:
- credential_exists
- data_node_exists
For querying if a service exists on the device. This API does not prompt
the user on the device.

This is used in SSH keys management to handle the case where
data_node "Moolticute SSH Keys" (first time) is not present on the
device. In that case no external loading dans decoding is started.