Implemented issue #253 Feature request: query user to input login if field is empty
If there is a new credential in the browser without login name, then daemon sends a signal to GUI for request the login from the user. When the user added his/her login name in the MessageBox, then the GUI sends a message to the daemon with the given login name.