Fixing crash on Mac from orderFrontRegardless
orderFrontRegardless was responsible on Mac to show the MainWindow on the top after it was hidden. However it crashed very often and even if it was not crashing, it hasn't brought the window in front.
It crashed because the used winId is changing during runtime, so it is dangerous to use it:
http://doc.qt.io/archives/qt-4.8/qwidget.html#winId
Instead of orderFrontRegardless the raise and activateWIndow call is bringing the MainWindow to the top level after clicking show without a crash.