#383 Fixing BLE dev tab layout for Mac, remove concurrent usb write on Mac
Concurrent run elimination:
In this case starting usb write in a new thread is causing problem because the sent message is containing more packets, and we sending write one-by-one, but if a previous packet's thread is blocked for some reason, it is causing an invalid order of packets.
And I don't think it is a problem if we block daemon's thread for this really short time, the user will not experience anything. (For windows and linux it isn't set to run on other thread, thats why those were working right)