Fix displaying previous device informations
On device disconnect the serial number field is cleared.
However when we connect the other device the parameters only read after enter the pin finished, but the previously connected device's parameters are still represented in wsClient (gui side) so on the connect signal it is loading those outdated informations.
Fix: On device disconnect I reset these informations, so until enter the pin is presented it is displaying "None" on About tab.