Remove Makefile in repository root
This Makefile was only used for the Debian package. It is therefore
confusing if this file lives in the root of the repository as it hints
that using `make` is the canonical way to build moolticute. Moreover, it
duplicates information that is already in the qmake project files.

As debhelper can handle qmake projects, only minor changes need to be
made to `debian/rules`. These are
- Copy the main `.pro` file to the name of the current working directory
  as this is where qmake looks for the main `.pro` file.
- Add `69-mooltipass.rules` to `debian/moolticute.install` as the udev
  rules are not considered by `qmake`, since they need to be downloaded
  from the separate mooltipass-udev repository. The `.install` file is
  read by `dh_install` to figure out which files need to be installed
  separately from what `qmake` is already installing.