Allow silent install without application launch
Some users may want to install moolticute in an unattended silent
way, e.g. the following installs in silent mode skipping the creation of
a desktop icon:

    setupfile.exe /VERYSILENT /TASKS="!desktopicon"

However this will still run programs in the `RUN` section which executes
moolticute. This is undesired in an unattended setup, so we skip this
[on silent installs][1].

[1]: https://documentation.help/Inno-Setup/topic_runsection.htm#skipifsilent